% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rashomon_set.R
\name{aggregate_rashomon_profiles}
\alias{aggregate_rashomon_profiles}
\title{Finds the RashomonSet of models for a given dataset.}
\usage{
aggregate_rashomon_profiles(
  data,
  value,
  arm_cols,
  M,
  R,
  H,
  reg = 1,
  theta,
  bruteforce = FALSE,
  inactive = 0
)
}
\arguments{
\item{data}{A dataframe containing the column whose name you supply in value}

\item{value}{The column name of the y values in data}

\item{arm_cols}{A character vector containing the names of the arm columns}

\item{M}{The number of arms}

\item{R}{An integer (or vector) denoting the number of levels in each arm. If
this value is an integer, it assumes each arm has the same number of levels.}

\item{H}{The maximum number of pools allowed to be specified by a partition matrix}

\item{reg}{The value of the regularization penalty when computing the loss for
a pooling structure. Defaults to 1, so that each additional pool in a model adds 1 to the loss.}

\item{theta}{Threshold value to be present in the RashomonSet. Set to Inf if you want all poolings.}

\item{bruteforce}{Whether the RashomonSet should be found via bruteforce. Currently unimplemented,
defaults to FALSE.}

\item{inactive}{The level that denotes an inactive arm in data. Defaults to zero.}
}
\value{
A list of RashomonSet Objects that gives all of the partition matrices in the
Rashomon Set for a given theta.
}
\description{
Finds the RashomonSet of models for a given dataset.
}
