% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{assign_universal_label}
\alias{assign_universal_label}
\title{Assign universal policy labels to data}
\usage{
assign_universal_label(data, arm_cols)
}
\arguments{
\item{data}{Dataframe containing the treatment assignments}

\item{arm_cols}{A character vector containing the names of the arm columns}
}
\description{
Assigns policy labels to data for use in create_policies_from_data. It will
assign a unique id to each unique treatment combination present in data.
}
