% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\name{combine_pool_dictionaries}
\alias{combine_pool_dictionaries}
\title{Combines all of the dictionaries of a RashomonSet object to form one dictionary.}
\usage{
combine_pool_dictionaries(obj)
}
\arguments{
\item{obj}{A RashomonSet object}

\item{newdata}{Data that we want to predict given our model object. Must contain
column universal_labels from output of assign_universal_labels(), and these labels
must correspond to the same policies as in the data passed to aggregate_rashomon_profiles()
or find_rashomon_profile()}
}
\description{
Combines all of the dictionaries of a RashomonSet object to form one dictionary.
}
