% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{compute_penalization_loss}
\alias{compute_penalization_loss}
\title{Computes penalization loss for data (regularization parameter * number of pools)}
\usage{
compute_penalization_loss(sigma, R, reg)
}
\arguments{
\item{sigma}{Partition matrix for a given pooling structure}

\item{R}{A list (or integer) of the number of levels in each arm.}

\item{reg}{Regularization parameter that penalizes partitions with more pools}
}
\value{
Penalization loss
}
\description{
Computes penalization loss for data (regularization parameter * number of pools)
}
