% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{create_policies}
\alias{create_policies}
\title{Creates all factor combinations in a factorial design}
\usage{
create_policies(levels, arms = 0)
}
\arguments{
\item{levels}{A number indicating the number of levels for each arm, or a list indicating
the number of levels for each arm}

\item{arms}{The number of arms, if levels is an integer.}
}
\value{
A list of all factor combinations (stored as numeric vectors) for a given number of arms and
levels per arm.
}
\description{
Creates all factor combinations in a factorial design
}
