% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{make_rashomon_objects}
\alias{make_rashomon_objects}
\title{Makes output from aggregate_rashomon_partitions cleaner}
\usage{
make_rashomon_objects(rset)
}
\description{
Makes output from aggregate_rashomon_partitions cleaner
}
