% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{pool_means}
\alias{pool_means}
\title{Finds pool means given data, value, and pool label}
\usage{
pool_means(data, value)
}
\arguments{
\item{data}{Data.table containing the treatment assignments}

\item{value}{Name of value column observed for each observation, supplied by the user. Must be character.}
}
\value{
A collections::dict() object, where the key is an integer i corresponding
to the pool id and its value is the mean of that pool.
}
\description{
Finds pool means given data, value, and pool label
}
