% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{pools_to_data}
\alias{pools_to_data}
\title{Assigns pools to observations in data given a dictionary of pools.}
\usage{
pools_to_data(data, pools_dict)
}
\arguments{
\item{data}{Data.table containing the treatment assignments}

\item{pools_dict}{A collections::dict object from extract_pools() that
takes in a policy id and outputs the pool it is in.}
}
\value{
A list of the sums of the products of size k, where the (k+1)th element
denotes the kth sum. The 1st element is 1, for use in other functions for easy looping.
}
\description{
Assigns pools to observations in data given a dictionary of pools.
}
