% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\name{predict.RashomonSet}
\alias{predict.RashomonSet}
\title{Gernerates predictions from a model in the RashomonSet}
\usage{
\method{predict}{RashomonSet}(obj, data)
}
\arguments{
\item{obj}{A RashomonSet object}

\item{newdata}{A dataframe. Must contain column universal_labels from output of assign_universal_labels(), and these labels
must correspond to the same policies as in the data passed to aggregate_rashomon_profiles()
or find_rashomon_profile()}
}
\description{
Gernerates predictions from a model in the RashomonSet
}
