% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomontva-package.R
\docType{package}
\name{rashomontva-package}
\alias{rashomontva}
\alias{rashomontva-package}
\title{rashomontva: Finding Rashomon Sets For Factorial Design Experiments}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Garrett Allen \email{gpa5@uw.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
